tests: Move socket.io connection helpers to `common.js`
